chore: rename bundles
