+++ packages/playback/rollup.config.js
@@ -11,7 +11,7 @@ export default [
   ...builder
     .copy()
     .setName('PlaybackNamespace')
-    .setInput('./src/entry-points/core-main.ts')
+    .setInput('./src/entry-points/main-core.ts')
     .setFolder('main/core')
     .build(),
 
@@ -19,8 +19,8 @@ export default [
   ...builder
     .copy()
     .setName('PlaybackNamespace')
-    .setInput('./src/entry-points/core-worker.ts')
-    .setWorker('dist/worker/bridge/es')
-    .setFolder('worker/core')
+    .setInput('./src/entry-points/main-with-worker-core.ts')
+    .setWorker('dist/worker/es')
+    .setFolder('main-with-worker/core')
     .build(),
 ];
