+++ packages/playback/src/entry-points/api-reference.ts
@@ -1,0 +1,23 @@
+/**
+ * this entry point is used for api reference generation via typedoc
+ */
+
+export type * from '../lib/types/mappers/event-type-to-event-map.declarations';
+export type * from '../lib/types/mappers/interceptor-type-to-interceptor-map.declarations';
+export type * from '../lib/types/audio-track.declarations';
+export type * from '../lib/types/configuration.declarations';
+export type * from '../lib/types/event-emitter.declarations';
+export type * from '../lib/types/interceptors.declarations';
+export type * from '../lib/types/logger.declarations';
+export type * from '../lib/types/mse-manager.declarations';
+export type * from '../lib/types/network.declarations';
+export type * from '../lib/types/pipeline.declarations';
+export type * from '../lib/types/playback-stats.declarations';
+export type * from '../lib/types/player-time-range.declarations';
+export type * from '../lib/types/quality-level.declarations';
+export type * from '../lib/types/retry.declarations';
+export type * from '../lib/types/source.declarations';
+export type * from '../lib/types/store.declarations';
+export type * from '../lib/types/text-track.declarations';
+export type * from '../lib/types/thumbnail-track.declarations';
+export type * from '../lib/types/utility.declarations';
