--- packages/playback/src/entry-points/core-main.ts
@@ -1,1 +1,0 @@
-export { Player } from '../lib/player/main-thread/player-main';
