--- packages/playback/src/entry-points/core-worker-bridge.ts
@@ -1,1 +1,0 @@
-export * from '../lib/player/worker-thread/worker-bridge';
