--- packages/playback/src/entry-points/core-worker.ts
@@ -1,1 +1,0 @@
-export { Player } from '../lib/player/worker-thread/player-worker';
