+++ packages/playback/src/entry-points/main-core.ts
@@ -1,0 +1,1 @@
+export { Player } from '../lib/player/main-thread-only/main-thread-only-player';
