+++ packages/playback/src/entry-points/main-with-worker-core.ts
@@ -1,0 +1,1 @@
+export { Player } from '../lib/player/main-thread-with-worker/main-thread-with-worker-player';
