+++ packages/playback/src/entry-points/worker-script.ts
@@ -1,0 +1,1 @@
+export * from '../lib/player/main-thread-with-worker/worker/worker-script';
