+++ packages/playback/src/lib/player/main-thread-only/main-thread-only-player.ts
